+++ CHANGES.md
@@ -11,6 +11,19 @@
 
 ## develop
 
+## 2021.3
+
+- [UPDATE] `cmake` を `3.20.2` に上げる
+    - @voluntas
+- [ADD] DataChannel シグナリングに対応
+    - @melpon
+- [ADD] WebSocket シグナリングの re-offer に対応
+    - @melpon
+- [FIX] サイマルキャストの active: false に対応
+    - @melpon
+
+## 2021.2
+
 - [UPDATE] `cmake` を `3.20.1` に上げる
     - @voluntas
 - [UPDATE] WebRTC のバージョンを M90 (4430@{#3}) に上げる